# public/openapi.yaml
openapi: 3.0.3
info:
  title: Fly2Any API
  description: |
    Comprehensive flight booking and management API with AI-powered search optimization,
    multi-source aggregation (Amadeus, Duffel), and intelligent caching.

    ## Features
    - Multi-airport and flexible date search
    - Real-time price confirmation
    - Seat selection and branded fares
    - Payment processing via Stripe
    - Booking management and modifications
    - Price alerts and saved searches

    ## Base URLs
    - Production: https://fly2any.com/api
    - Development: http://localhost:3000/api
  version: 1.0.0
  contact:
    name: Fly2Any API Support
    email: api-support@fly2any.com
    url: https://fly2any.com/docs
  license:
    name: Proprietary
    url: https://fly2any.com/terms
servers:
  - url: https://fly2any.com/api
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
...
security:
  - sessionAuth: []
paths:
...
  /payments/create-intent:
    post:
      tags:
        - Payments
      summary: Create payment intent
      description: Create a Stripe payment intent for booking payment
      operationId: createPaymentIntent
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentIntentRequest'
      responses:
        '201':
          description: Payment intent created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentIntentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /payments/confirm:
    post:
      tags:
        - Payments
      summary: Confirm payment
      description: Confirm payment after completion (including 3D Secure)
      operationId: confirmPayment
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - paymentIntentId
                - bookingReference
              properties:
                paymentIntentId:
                  type: string
                  description: Stripe payment intent ID
                bookingReference:
                  type: string
                  description: Booking reference number
      responses:
        '200':
          description: Payment confirmed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentConfirmationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: Session-based authentication via NextAuth.js
  schemas:
...
    CreatePaymentIntentRequest:
      type: object
      required:
        - amount
        - currency
        - customerEmail
        - customerName
      properties:
        amount:
          type: number
          minimum: 0
        currency:
          type: string
        bookingId:
          type: string
        bookingReference:
          type: string
        customerEmail:
          type: string
          format: email
        customerName:
          type: string
        description:
          type: string
        metadata:
          type: object
    PaymentIntentResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            paymentIntent:
              type: object
              properties:
                paymentIntentId:
                  type: string
                clientSecret:
                  type: string
                amount:
                  type: number
                currency:
                  type: string
                status:
                  type: string
            bookingReference:
              type: string
        meta:
          type: object
          properties:
            timestamp:
              type: string
              format: date-time
            processingTime:
              type: integer
    PaymentConfirmationResponse:
      type: object
      properties:
        success:
          type: boolean
        payment:
          type: object
          properties:
            paymentIntentId:
              type: string
            status:
              type: string
            amount:
              type: number
            currency:
              type: string
            paymentMethod:
              type: string
            last4:
              type: string
            brand:
              type: string
        booking:
          type: object
          properties:
            id:
              type: string
            bookingReference:
              type: string
            status:
              type: string
        message:
          type: string
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          default: false
        error:
          oneOf:
            - type: string
            - type: object
              properties:
                code:
                  type: string
                message:
                  type: string
                details:
                  type: object
        meta:
          type: object
          properties:
            timestamp:
              type: string
              format: date-time
  responses:
    BadRequest:
      description: Bad request - invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...